// frontend/src/app/core/api/generated/api/inventory.service.ts
export interface FindAll10RequestParams {
    page?: number;
    size?: number;
}