// frontend/src/app/core/api/generated/api/warehouse-inventory.service.ts
export interface FindSettingsRequestParams {
    warehouseId: string;
    page?: number;
    size?: number;
}