// frontend/src/app/core/api/generated/model/inventory-setting-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InventorySettingResponse { 
    warehouseId?: string;
    productId?: string;
    sku?: string;
    name?: string;
    minimumStock?: number;
    active?: boolean;
}